make bcache tests run

+++ tests/bcache/bcache-test-libs.sh
@@ -8,6 +8,9 @@ require-build-deb bcache-tools
 
 require-kernel-config MD
 require-kernel-config BCACHE,BCACHE_DEBUG
+require-kernel-config AUTOFS_FS
+
+require-make ../../ltp-fsx/
 
 if [[ $KERNEL_ARCH = x86 ]]; then
     require-kernel-config CRYPTO_CRC32C_INTEL
@@ -25,6 +28,7 @@ WRITEAROUND=0
 REPLACEMENT=lru
 BUCKET_SIZE=""
 BLOCK_SIZE=""
+TEST_RUNNING=""
 
 VIRTIO_BLKDEVS=0
 
@@ -39,6 +43,11 @@ config-backing()
 config-cache()
 {
     add_bcache_devs CACHE $1 0
+}
+
+config-tier()
+{
+    add_bcache_devs TIER $1 1
 }
 
 config-volume()
@@ -112,7 +121,13 @@ add_bcache_devs()
 
 bcache_format()
 {
-    make-bcache $BUCKET_SIZE $BLOCK_SIZE -C $CACHE -B $BDEV
+    for dev in ${CACHE} ${BDEV}
+    do
+        wipefs -a ${dev}
+    done
+
+    make_bcache_flags=" --wipe-bcache"
+    make-bcache $BUCKET_SIZE $BLOCK_SIZE -C $CACHE -B $BDEV ${make_bcache_flags}
 }
 
 add_device() {
@@ -142,7 +157,9 @@ existing_bcache() {
     echo "registering via sysfs"
 
     for dev in $CACHE $BDEV; do
-	echo $dev > /sys/fs/bcache/register
+	if ! echo $dev > /sys/fs/bcache/register; then
+            echo "WARN" "${dev} maybe have been registered by systemd with udev"
+	fi
     done
 
     echo "registered"
@@ -184,13 +201,15 @@ setup_bcache() {
     sleep 2
 
     for size in $VOLUME; do
-	for file in /sys/fs/bcache/*/blockdev_volume_create; do
+	for file in /sys/fs/bcache/*/flash_vol_create; do
 	    echo "creating volume $size via $file"
 	    echo $size > $file
 	done
     done
 
-    ln -s /sys/fs/bcache/*-* /root/c || true
+    if [ ! -L "/root/c" ]; then
+        ln -s /sys/fs/bcache/*-* /root/c || true
+    fi
 }
 
 stop_volumes()
@@ -350,10 +369,10 @@ discard_all_devices()
     fi
 
     expect_sysfs cache dirty_buckets 0
-    expect_sysfs cache dirty_data 0
+    expect_sysfs cache dirty_data 0.0k
     expect_sysfs cache cached_buckets 0
     expect_sysfs cache cached_data 0
-    expect_sysfs bdev dirty_data 0
+    expect_sysfs bdev dirty_data 0.0k
 
     if [ -f /sys/kernel/debug/bcache/* ]; then
 	tmp="$(mktemp)"
@@ -394,4 +413,185 @@ run_bcache_stress()
     fi
 
     disable_faults
+}
+
+block_device_verify_dd()
+{
+    dd if=$1 of=/root/cmp bs=4096 count=1 iflag=direct
+    cmp /root/cmp /root/orig
+}
+
+block_device_dd()
+{
+    dd if=/dev/urandom of=/root/orig bs=4096 count=1
+    dd if=/root/orig of=$1 bs=4096 count=1 oflag=direct
+    dd if=$1 of=/root/cmp bs=4096 count=1 iflag=direct
+    cmp /root/cmp /root/orig
+
+    dd if=/dev/urandom of=/root/orig bs=4096 count=1
+    dd if=/root/orig of=$1 bs=4096 count=1 oflag=direct
+}
+
+wait_all()
+{
+    for job in $(jobs -p); do
+        wait $job
+    done
+}
+
+run_fio()
+{
+    echo "=== start fio at $(date)"
+    loops=$(($ktest_priority / 2 + 1))
+
+    (
+        # Our default working directory (/cdrom) is not writable,
+        # fio wants to write files when verify_dump is set, so
+        # change to a different directory.
+        cd $LOGDIR
+
+        for dev in $DEVICES; do
+            fio --eta=always            \
+                --randrepeat=0          \
+                --ioengine=libaio       \
+                --iodepth=64            \
+                --iodepth_batch=16      \
+                --direct=1              \
+                --numjobs=1             \
+                --buffer_compress_percentage=20\
+                --verify=meta           \
+                --verify_fatal=1        \
+                --verify_dump=1         \
+                --filename=$dev         \
+                --fill_fs=1             \
+                                        \
+                --name=seqwrite         \
+                --stonewall             \
+                --rw=write              \
+                --bsrange=4k-128k       \
+                --loops=$loops          \
+                                        \
+                --name=randwrite        \
+                --stonewall             \
+                --rw=randwrite          \
+                --bsrange=4k-128k       \
+                --loops=$loops          \
+                                        \
+                --name=randwrite_small  \
+                --stonewall             \
+                --rw=randwrite          \
+                --bs=4k                 \
+                --loops=$loops          \
+                                        \
+                --name=randread         \
+                --stonewall             \
+                --rw=randread           \
+                --bs=4k                 \
+                --loops=$loops          &
+        done
+
+        wait_all
+    )
+
+    echo "=== done fio at $(date)"
+}
+
+#
+# Mount file systems on all block devices.
+#
+existing_fs() {
+    case $1 in
+       ext4)
+           opts="-o errors=panic"
+           ;;
+       xfs)
+           opts=""
+           ;;
+       *)
+           opts=""
+           ;;
+    esac
+
+    for dev in $DEVICES; do
+       mkdir -p /mnt/$dev
+       mount $dev /mnt/$dev -t $1 $opts
+    done
+}
+
+#
+# Set up file systems on all block devices and mount them.
+#
+setup_fs()
+{
+    for dev in $DEVICES; do
+       case $1 in
+           xfs)
+               opts="-f"
+               ;;
+           ext4)
+               opts="-F"
+               ;;
+           *)
+               opts=""
+               ;;
+       esac
+
+       mkfs.$1 $opts $dev
+    done
+    existing_fs $1
+}
+
+run_dbench()
+{
+    echo "=== start dbench at $(date)"
+    duration=$((($ktest_priority + 1) * 30))
+
+    (
+        for dev in $DEVICES; do
+            dbench -S -t $duration 2 -D /mnt/$dev &
+        done
+
+        wait_all
+    )
+
+    echo "=== done dbench at $(date)"
+}
+
+run_bonnie()
+{
+    echo "=== start bonnie at $(date)"
+    loops=$((($ktest_priority + 1) * 4))
+
+    (
+        for dev in $DEVICES; do
+            bonnie++ -x $loops -r 128 -u root -d /mnt/$dev &
+        done
+
+        wait_all
+    )
+
+    echo "=== done bonnie at $(date)"
+}
+
+run_fsx()
+{
+    echo "=== start fsx at $(date)"
+    numops=$((($ktest_priority + 1) * 300000))
+
+    (
+        for dev in $DEVICES; do
+            ltp-fsx -N $numops /mnt/$dev/foo &
+        done
+
+        wait_all
+    )
+
+    echo "=== done fsx at $(date)"
+}
+
+stop_fs()
+{
+    for dev in $DEVICES; do
+        umount /mnt/$dev
+    done
 }
